Consolidate Hermes adapter package exports
