+++ packages/adapters/hermes-gateway/src/cli/index.ts
@@ -1,2 +1,1 @@
-export { printHermesGatewayStreamEvent as formatStdoutEvent } from "./format-event.js";
-
+export { formatStdoutEvent } from "@paperclipai/hermes-paperclip-adapter/gateway/cli";
