+++ packages/adapters/hermes/package.json
@@ -1,19 +1,19 @@
 {
   "name": "@paperclipai/hermes-paperclip-adapter",
   "version": "0.3.1",
-  "description": "Paperclip adapter for Hermes Agent — run Hermes as a managed employee in a Paperclip company",
+  "description": "Paperclip adapters for Hermes Agent local CLI and Hermes Gateway HTTP/SSE runs",
   "type": "module",
   "license": "MIT",
   "author": "Paperclip",
   "repository": {
     "type": "git",
     "url": "https://github.com/paperclipai/paperclip",
-    "directory": "packages/adapters/hermes-local"
+    "directory": "packages/adapters/hermes"
   },
   "bugs": {
     "url": "https://github.com/paperclipai/paperclip/issues"
   },
-  "homepage": "https://github.com/paperclipai/paperclip/tree/master/packages/adapters/hermes-local#readme",
+  "homepage": "https://github.com/paperclipai/paperclip/tree/master/packages/adapters/hermes#readme",
   "keywords": [
     "paperclip",
     "hermes",
@@ -27,7 +27,12 @@
     "./server": "./src/server/index.ts",
     "./ui": "./src/ui/index.ts",
     "./cli": "./src/cli/index.ts",
-    "./ui-parser": "./ui-parser.cjs"
+    "./ui-parser": "./ui-parser.cjs",
+    "./gateway": "./src/gateway/index.ts",
+    "./gateway/server": "./src/gateway/server/index.ts",
+    "./gateway/ui": "./src/gateway/ui/index.ts",
+    "./gateway/cli": "./src/gateway/cli/index.ts",
+    "./gateway/ui-parser": "./gateway-ui-parser.cjs"
   },
   "paperclip": {
     "adapterUiParser": "1.0.0"
@@ -51,7 +56,24 @@
         "types": "./dist/cli/index.d.ts",
         "import": "./dist/cli/index.js"
       },
-      "./ui-parser": "./ui-parser.cjs"
+      "./ui-parser": "./ui-parser.cjs",
+      "./gateway": {
+        "types": "./dist/gateway/index.d.ts",
+        "import": "./dist/gateway/index.js"
+      },
+      "./gateway/server": {
+        "types": "./dist/gateway/server/index.d.ts",
+        "import": "./dist/gateway/server/index.js"
+      },
+      "./gateway/ui": {
+        "types": "./dist/gateway/ui/index.d.ts",
+        "import": "./dist/gateway/ui/index.js"
+      },
+      "./gateway/cli": {
+        "types": "./dist/gateway/cli/index.d.ts",
+        "import": "./dist/gateway/cli/index.js"
+      },
+      "./gateway/ui-parser": "./gateway-ui-parser.cjs"
     },
     "main": "./dist/index.js",
     "types": "./dist/index.d.ts",
@@ -61,6 +83,7 @@
     "dist",
     "skills",
     "ui-parser.cjs",
+    "gateway-ui-parser.cjs",
     "README.md",
     "LICENSE"
   ],
@@ -81,7 +104,8 @@
   },
   "devDependencies": {
     "@types/node": "^22.19.21",
-    "typescript": "^5.7.3"
+    "typescript": "^5.7.3",
+    "vitest": "^4.1.8"
   },
   "engines": {
     "node": ">=20.0.0"
