+++ packages/adapters/hermes/scripts/prepare-publish-package.mjs
