+++ packages/adapters/hermes/skills/paperclip-task-bridge/SKILL.md
