+++ packages/adapters/hermes/skills/paperclip-task-bridge/paperclip-task.mjs
