+++ packages/adapters/hermes/src/cli/format-event.ts
