+++ packages/adapters/hermes/src/cli/index.ts
