+++ packages/adapters/hermes/src/gateway/cli/format-event.ts
