+++ packages/adapters/hermes/src/gateway/cli/index.ts
@@ -1,0 +1,2 @@
+export { printHermesGatewayStreamEvent as formatStdoutEvent } from "./format-event.js";
+
