+++ packages/adapters/hermes/src/gateway/index.ts
@@ -1,0 +1,73 @@
+import type { AdapterSessionManagement, ServerAdapterModule } from "@paperclipai/adapter-utils";
+import { ADAPTER_LABEL, ADAPTER_TYPE } from "./shared/constants.js";
+import { execute, getConfigSchema, sessionCodec, testEnvironment } from "./server/index.js";
+
+export const type = ADAPTER_TYPE;
+export const label = ADAPTER_LABEL;
+export const models: { id: string; label: string }[] = [];
+
+const sessionManagement: AdapterSessionManagement = {
+  supportsSessionResume: true,
+  nativeContextManagement: "confirmed",
+  defaultSessionCompaction: {
+    enabled: true,
+    maxSessionRuns: 0,
+    maxRawInputTokens: 0,
+    maxSessionAgeHours: 0,
+  },
+};
+
+export const agentConfigurationDoc = `# hermes_gateway agent configuration
+
+Adapter: hermes_gateway
+
+Use when:
+- Hermes Agent runs on another host or process that exposes the Hermes API server.
+- Paperclip should create Hermes runs through POST /v1/runs and observe them through SSE events.
+- You need remote Hermes session continuity with X-Hermes-Session-Key.
+
+Don't use when:
+- Hermes should run as a local child process on the Paperclip host; use hermes_local instead.
+- The Hermes API server is not enabled or is only reachable over an unsafe public HTTP endpoint.
+- You need Hermes-originated Paperclip task creation only; that is a Paperclip skill/API bridge, not this wake adapter.
+
+Required fields:
+- apiBaseUrl (string): Hermes API server base URL, for example http://127.0.0.1:8642.
+- apiKey (string): Hermes API_SERVER_KEY. Sent as Authorization: Bearer <apiKey>.
+
+Optional fields:
+- headers (object or JSON string): extra noncritical headers. Authorization, Content-Type, Accept, Idempotency-Key, and X-Hermes-Session-Key are generated by the adapter.
+- paperclipApiUrl (string): Paperclip API URL reachable from the Hermes host. This is not a credential.
+- sessionKeyStrategy (issue | agent | run | none): defaults to issue.
+- timeoutSec (number): defaults to 120.
+- eventReconnectMs (number): defaults to 2000.
+- instructions (string): stable Hermes instructions sent separately from wake input.
+
+Runtime mapping:
+- Creates runs with POST /v1/runs.
+- Sends Idempotency-Key equal to the Paperclip run id for correlation only; Hermes v0.16.0 did not dedupe duplicate creates.
+- Streams GET /v1/runs/{run_id}/events and polls GET /v1/runs/{run_id} as fallback.
+- Calls POST /v1/runs/{run_id}/stop on timeout.
+
+Security guidance:
+- Prefer HTTPS or a private overlay network for non-loopback hosts.
+- Do not put Hermes apiKey or Paperclip bearer tokens in prompts, comments, logs, or result JSON.
+- Keep the default issue-scoped session strategy unless shared agent memory is intentional.
+`;
+
+export function createServerAdapter(): ServerAdapterModule {
+  return {
+    type,
+    execute,
+    testEnvironment,
+    sessionCodec,
+    sessionManagement,
+    models,
+    supportsLocalAgentJwt: false,
+    supportsInstructionsBundle: false,
+    requiresMaterializedRuntimeSkills: false,
+    agentConfigurationDoc,
+    getConfigSchema,
+  };
+}
+
