+++ packages/adapters/hermes/src/gateway/server/config-schema.ts
