+++ packages/adapters/hermes/src/gateway/server/transport-security.ts
