+++ packages/adapters/hermes/src/gateway/shared/constants.ts
