+++ packages/adapters/hermes/src/gateway/ui/index.ts
@@ -1,0 +1,2 @@
+export { parseHermesGatewayStdoutLine as parseStdoutLine } from "./parse-stdout.js";
+
