+++ packages/adapters/hermes/src/gateway/ui/parse-stdout.ts
