+++ packages/adapters/hermes/src/server/command-resolution.test.ts
