+++ packages/adapters/hermes/src/server/config-schema.ts
