+++ packages/adapters/hermes/src/server/detect-model.test.ts
