+++ packages/adapters/hermes/src/server/detect-model.ts
