+++ packages/adapters/hermes/src/server/prompt-rendering.test.ts
