+++ packages/adapters/hermes/src/shared/constants.ts
