+++ packages/adapters/hermes/src/ui/parse-stdout.ts
