+++ scripts/bootstrap-npm-package.test.mjs
@@ -60,11 +60,11 @@ test("resolveTargetPackage includes the workspace diff plugin bootstrap package"
 });
 
 test("buildPublishArgs publishes from the repo root through pnpm", () => {
-  const pkg = { dir: "packages/adapters/hermes-local", name: "@paperclipai/hermes-paperclip-adapter" };
+  const pkg = { dir: "packages/adapters/hermes", name: "@paperclipai/hermes-paperclip-adapter" };
 
   assert.deepEqual(buildPublishArgs(pkg), [
     "publish",
-    "packages/adapters/hermes-local",
+    "packages/adapters/hermes",
     "--no-git-checks",
     "--access",
     "public",
@@ -72,11 +72,11 @@ test("buildPublishArgs publishes from the repo root through pnpm", () => {
 });
 
 test("buildPublishArgs includes dry-run and otp flags when requested", () => {
-  const pkg = { dir: "packages/adapters/hermes-local", name: "@paperclipai/hermes-paperclip-adapter" };
+  const pkg = { dir: "packages/adapters/hermes", name: "@paperclipai/hermes-paperclip-adapter" };
 
   assert.deepEqual(buildPublishArgs(pkg, { dryRun: true, otp: "123456" }), [
     "publish",
-    "packages/adapters/hermes-local",
+    "packages/adapters/hermes",
     "--no-git-checks",
     "--access",
     "public",
