+++ scripts/generate-npm-package-json.mjs
@@ -33,7 +33,7 @@ const workspacePaths = [
   "packages/adapters/claude-local",
   "packages/adapters/codex-local",
   "packages/adapters/hermes-gateway",
-  "packages/adapters/hermes-local",
+  "packages/adapters/hermes",
   "packages/adapters/opencode-local",
   "packages/adapters/openclaw-gateway",
 ];
